Refactor OData handler data loader
