+++ packages/handlers/odata/src/util.ts
@@ -1,3 +1,4 @@
+import { ResolverData } from "@graphql-mesh/types";
 
 export function getUrlString(url: URL) {
   return decodeURIComponent(url.toString()).split('+').join(' ');
